--- controller/api/metrics.py
@@ -1,6 +1,0 @@
-"""
-metrics.py
-"""
-from fastapi import APIRouter
-
-router = APIRouter(prefix="/api/metrics", tags=["metrics"])
