+++ settings.py
@@ -39,7 +39,7 @@ class Settings(BaseSettings):
     syslog_port: int = 514
 
     # Default task directory.
-    task_dir: DirectoryPath = "controller/worker/tasks"
+    task_dir: DirectoryPath = "worker_service/tasks"
 
     class Config:
         """Modify the behaviour of settings management"""
